+++ src/nature/OnePlusOneBitString.java
@@ -21,8 +21,8 @@ public class OnePlusOneBitString extends Algorithm<BitString> {
 	public void step(long iteration) {
 		BitString mutation = current.globalMutation(globalMutationProbability);
 
-		double currentFitness = fitnessGoal.evaluate(current);
-		double mutationFitness = fitnessGoal.evaluate(mutation);
+		int currentFitness = fitnessGoal.evaluate(current);
+		int mutationFitness = fitnessGoal.evaluate(mutation);
 
 		if (fitnessGoal.compare(currentFitness, mutationFitness) <= 0) {
 			current = mutation;
